# jackdaw/nest/api/swagger.yaml
swagger: "2.0"
info:
  description: This is the swagger file that goes with our server code
  version: "1.0.0"
  title: Swagger Rest Article
consumes:
  - application/json
produces:
  - application/json
basePath: /
...
paths:
...
  /scans/enum/create:
    post:
      operationId: jackdaw.nest.api.scan.scan_enum
      tags:
        - Scan
      summary: Schedules a new enumeration scan
      description: Schedules a new enumeration scan
      parameters:
        - name: params
          in: body
          description: scan parameters
          schema:
            type: object
            properties:
              ldap_url:
                type: string
              smb_url:
                type: string
              ldap_workers:
                type: integer
              smb_workers:
                type: integer
      responses:
        200:
          description: Nah
          schema:
            type: object
            properties:
              graph:
                type: string
                description: Placeholder TODO 
...